processor/requirements.txt
flask
exceptiongroup
langchain==0.0.232
openai
pinecone-client
unstructured
psycopg2-binary==2.9.1